feat: user can update their data

+++ src/main/java/com/github/skyg0d/skydrinksapi/service/ApplicationUserService.java
@@ -1,6 +1,8 @@
 package com.github.skyg0d.skydrinksapi.service;
 
 import com.github.skyg0d.skydrinksapi.domain.ApplicationUser;
+import com.github.skyg0d.skydrinksapi.enums.Roles;
+import com.github.skyg0d.skydrinksapi.exception.ActionNotAllowedException;
 import com.github.skyg0d.skydrinksapi.exception.BadRequestException;
 import com.github.skyg0d.skydrinksapi.exception.UserUniqueFieldExistsException;
 import com.github.skyg0d.skydrinksapi.mapper.ApplicationUserMapper;
@@ -10,6 +12,7 @@ import com.github.skyg0d.skydrinksapi.repository.user.ApplicationUserSpecificati
 import com.github.skyg0d.skydrinksapi.requests.ApplicationUserPostRequestBody;
 import com.github.skyg0d.skydrinksapi.requests.ApplicationUserPutRequestBody;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -19,6 +22,7 @@ import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
+@Log4j2
 public class ApplicationUserService {
 
     private final ApplicationUserRepository applicationUserRepository;
@@ -72,13 +76,25 @@ public class ApplicationUserService {
         return applicationUserRepository.save(mapper.toApplicationUser(applicationUserPostRequestBody));
     }
 
-    public void replace(ApplicationUserPutRequestBody applicationUserPutRequestBody) {
-        findByIdOrElseThrowBadRequestException(applicationUserPutRequestBody.getUuid());
+    public void replace(ApplicationUserPutRequestBody applicationUserPutRequestBody, ApplicationUser user) {
+        UUID uuid = applicationUserPutRequestBody.getUuid();
+
+        verifyIfUserHasPermission(uuid, user);
+
+        findByIdOrElseThrowBadRequestException(uuid);
+
         applicationUserRepository.save(mapper.toApplicationUser(applicationUserPutRequestBody));
     }
 
-    public void delete(UUID uuid) {
+    public void delete(UUID uuid, ApplicationUser user) {
+        verifyIfUserHasPermission(uuid, user);
         applicationUserRepository.delete(findByIdOrElseThrowBadRequestException(uuid));
+    }
+
+    private void verifyIfUserHasPermission(UUID uuid, ApplicationUser user) {
+        if (!user.getRole().contains(Roles.ADMIN.getName()) && !user.getUuid().equals(uuid)) {
+            throw new ActionNotAllowedException("Apenas o usuário original ou admins podem alterar dados.");
+        }
     }
 
 }
